backend/requirements.txt
fastapi
uvicorn
httpx
python-dotenv
pgvector
psycopg2-binary
sentence-transformers
